Break down server index into smaller components of app, routes, and controllers
