+++ src/server/routes/api.ts
@@ -1,0 +1,8 @@
+import { Router } from 'express';
+import { proxyController } from '../controllers/apiControllers';
+
+const router = Router();
+
+router.get('*', proxyController);
+
+export default router;
