+++ src/server/routes/ui.ts
@@ -1,0 +1,8 @@
+import { Router } from 'express';
+import uiRootController from '../controllers/uiControllers';
+
+const router = Router();
+
+router.get('*', uiRootController);
+
+export default router;
